refacto: changed type of width and height in game struct

+++ server/include/game/game.h
@@ -10,8 +10,8 @@
 #include "map/map.h"
 
 typedef struct s_game {
-    int height;
-    int width;
+    size_t height;
+    size_t width;
     tile_t **map;
 } game_t;
